Fix one specific case of empty comments
This is a bit of a hack to avoid parsing every comments markdown twice.
Currently the final text is rendered in the result step, so we have
no way of knowing if a comment is empty at this point in time.
However the only instance of empty comments that I found so far were
/r/anime comments that had an embedded image using css trickery.

These comments are of the format [](/some_image_name).